% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGMissingDataR-package.R
\docType{package}
\name{CGMissingDataR-package}
\alias{CGMissingDataR}
\alias{CGMissingDataR-package}
\title{CGMmissingDataR: Impute Missing Glucose Values in CGM Data}
\description{
Imputes missing glucose values in repeated-measures continuous glucose
monitoring (CGM) data. Workflows create time-series features from raw
timestamps, support model selection, and return model-specific completed
data sets for glucose values that are already missing in user data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zhanglabuky.github.io/CGMmissingDataR/}
  \item \url{https://github.com/ZhangLabUKY/CGMmissingDataR}
  \item Report bugs at \url{https://github.com/ZhangLabUKY/CGMmissingDataR/issues}
}

}
\author{
\strong{Maintainer}: Shubh Saraswat \email{shubh.saraswat00@gmail.com} (\href{https://orcid.org/0009-0009-2359-1484}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Hasin Shahed Shad \email{hasin.shad@uky.edu}
  \item Xiaohua Douglas Zhang \email{douglas.zhang@uky.edu} (\href{https://orcid.org/0000-0002-2486-7931}{ORCID})
}

}
\keyword{internal}
